requirements.txt
Flask==2.3.3
Werkzeug==2.3.7
mysql-connector-python==8.1.0
gunicorn==21.2.0
python-dotenv==1.0.0
